+++ npm-search/src/js/app.js
@@ -1,6 +1,6 @@
 var React           = window.React = require('react'), // assign it to window for react chrome extension
-    classNames      = require('classnames'),
     SearchBar       = require('./searchbar.jsx'),
+    PluginList      = require('./pluginlist.jsx'),
     App             = {};
 
 var Constants = {
@@ -8,255 +8,12 @@ var Constants = {
     NpmSearchInitialSize: 500
 }
 
-var PlatformButton = React.createClass({
-    onClick: function() {
-        var appInstance = React.render(<App />, document.getElementById('container'));
-        appInstance.addCondition("platform:" + this.props.platform);
-    },
-    render: function() {
-        return (
-            <li className = "clickable" onClick={this.onClick}> {this.props.platform} </li>
-        );
-    }
-});
-
-var SupportedPlatforms = React.createClass({
-    getInitialState: function() {
-        return {
-            moreClicked: false
-        };
-    },
-    onClick: function() {
-        this.setState({
-            moreClicked: true
-        });
-    },
-    render: function() {
-        var keywords = this.props.keywords;
-        var sortedMajorPlatforms = [{present:false, text: "Android"}, {present:false, text: "iOS"}, {present:false, text: "Windows"}, {present:false, text: "Blackberry10"}];
-        var majorPlatformsSupported = [];
-        var otherPlatformsSupported = [];
-        // remove windows8 & windows dupe
-        if (keywords.indexOf('cordova-windows') > -1 && keywords.indexOf('cordova-windows8') > -1) {
-            keywords.splice(keywords.indexOf('cordova-windows8'), 1);
-        }
-        keywords.forEach(function(keyword) {
-            switch (keyword) {
-                case 'cordova-firefoxos':
-                    otherPlatformsSupported.push(<PlatformButton platform="FirefoxOS" />);
-                    break;
-                case 'cordova-android':
-                    sortedMajorPlatforms[0].present = true;
-                    break;
-                case 'cordova-amazon-fireos':
-                    otherPlatformsSupported.push(<PlatformButton platform="FireOS" />);
-                    break;
-                case 'cordova-ubuntu':
-                    otherPlatformsSupported.push(<PlatformButton platform="Ubuntu" />);
-                    break;
-                case 'cordova-ios':
-                    sortedMajorPlatforms[1].present = true;
-                    break;
-                case 'cordova-blackberry10':
-                    sortedMajorPlatforms[3].present = true;
-                    break;
-                case 'cordova-wp8':
-                    otherPlatformsSupported.push(<PlatformButton platform="Windows Phone 8" />);
-                    break;
-                case 'cordova-windows8':
-                case 'cordova-windows':
-                    sortedMajorPlatforms[2].present = true;
-                    break;
-                case 'cordova-browser':
-                    otherPlatformsSupported.push(<PlatformButton platform="Browser" />);
-                    break;
-            }
-        });
-
-        sortedMajorPlatforms.forEach(function(platform) {
-            if(platform.present) {
-                majorPlatformsSupported.push(<PlatformButton platform={platform.text} />)
-            }
-        });
-        while(majorPlatformsSupported.length < 4 && otherPlatformsSupported.length > 0) {
-            majorPlatformsSupported.push(otherPlatformsSupported.shift());
-        }
-
-        var moreButton;
-        if(otherPlatformsSupported.length > 0 && !this.state.moreClicked) {
-            moreButton = <li className="clickable" onClick={this.onClick}>...</li>
-        }
-        if (!this.state.moreClicked) {
-            otherPlatformsSupported = null;
-        }
-        return (
-            <ul className="supportedPlatforms">
-                {majorPlatformsSupported}
-                {moreButton}
-                {otherPlatformsSupported}
-            </ul>
-        );
-    }
-});
-
-var Plugin = React.createClass({
-    shouldComponentUpdate: function(nextProps, nextState) {
-        return this.props.plugin !== nextProps.plugin;
-    },
-    render: function() {
-        var license = this.props.plugin.license;
-        if (license && license.length > 1) {
-            license = license[0];
-        }
-        var downloadField;
-
-        var classes = classNames({
-            'pluginCard': true,
-            'featuredPlugin': this.props.plugin.isOfficial,
-            'row': true
-        });
-
-        if(this.props.plugin.downloadCount) {
-            var downloadCount = this.props.plugin.downloadCount.toLocaleString();
-            downloadField = <p><small> {downloadCount} downloads last month</small></p>;
-        }
-        return (
-            <li>
-                <div className={classes}>
-                    <div className="primaryContent col-xs-9">
-                        <div className="header">
-                            <h3><a href={'https://www.npmjs.com/package/' + this.props.plugin.name} target="_blank">{this.props.plugin.name}</a></h3>
-                            <small className="pluginVersion">v{this.props.plugin.version}</small>
-                            <small> by </small>
-                            <small className="pluginAuthor">{this.props.plugin.author}</small>
-                        </div>
-                        <div className="pluginDesc">{this.props.plugin.description}</div>
-                        <SupportedPlatforms keywords={this.props.plugin.keywords}/>
-                        </div>
-                        <div className="secondaryContent col-xs-3">
-                        <div className="download">
-                            <p></p>
-                        </div>
-                        <div className="extraInfo">
-                            <p><small><strong>License:</strong> {license}</small></p>
-                            {downloadField}
-                            <p><small><em>Last updated {this.props.plugin.modified} days ago</em></small></p>
-                        </div>
-                    </div>
-                </div>
-            </li>
-        )
-    }
-});
-
-var timer=null;
-
 window.addEventListener('popstate', function(e) {
     if(e.state) {
         var appInstance = React.render(<App />, document.getElementById('container'));
         appInstance.loadFilterText(e.state.filterText);
     }
 });
-
-var PluginList = React.createClass({
-    statics: {
-        contains: function(values, pluginInfo) {
-            var allValuesPresent = true;
-            if(values.length == 0) {
-                return allValuesPresent;
-            }
-            if(!pluginInfo) {
-                return false;
-            }
-            values.forEach(function(value) {
-                var valuePresent = false;
-                for(var index=0; index < pluginInfo.length; index++) {
-                    if(pluginInfo[index] && pluginInfo[index].toLowerCase().indexOf(value) > -1) {
-                        valuePresent = true;
-                    }
-                }
-                if(!valuePresent) {
-                    allValuesPresent = false;
-                }
-            });
-            return allValuesPresent;
-        },
-        populateFilters: function(filterText)
-        {
-            var searchStrings = filterText.split(" ");
-            var filters = [];
-            filters['platforms'] = [];
-            filters['authors'] = [];
-            filters['licenses'] = [];
-            filters['searchWords'] = [];
-
-            searchStrings.forEach(function(searchString) {
-                var keywords = searchString.split(":");
-                if(keywords.length == 1) {
-                    var param = keywords[0].trim();
-                    if(param) {
-                        filters['searchWords'].push(param);
-                    }
-                }
-                else if(keywords[1].trim()) {
-                    var param = keywords[1].trim();
-                    switch(keywords[0]) {
-                        case 'platform':
-                            filters['platforms'].push(param);
-                            break;
-                        case 'author':
-                            filters['authors'].push(param);
-                            break;
-                        case 'license':
-                            filters['licenses'].push(param);
-                            break;
-                        default:
-                            filters['searchWords'].push(searchString);
-                    }
-                }
-                else {
-                    filters['searchWords'].push(searchString);
-                }
-            });
-            return filters;
-        }
-    },
-    render: function() {
-        var plugins = [],
-            filterText = this.props.filterText.toLowerCase();
-
-        var delay = (function(){
-          return function(callback, ms){
-            clearTimeout (timer);
-            timer = setTimeout(callback, ms);
-          };
-        })();
-
-        delay(function(){
-                    window.history.pushState({"filterText":filterText}, "", "?q=" + filterText);
-            }, 2000 );
-
-        var filters = PluginList.populateFilters(filterText);
-
-        this.props.plugins.forEach(function(plugin) {
-            var fullPluginText = plugin.name.concat(plugin.author, plugin.keywords, plugin.license, plugin.description);
-            if(PluginList.contains(filters['platforms'], plugin.keywords)
-                && PluginList.contains(filters['authors'], plugin.author)
-                && PluginList.contains(filters['licenses'], plugin.license)
-                && PluginList.contains(filters['searchWords'], fullPluginText)) {
-                    plugins.push(<Plugin plugin={plugin} key={plugin.author + plugin.name}/>);
-            }
-        }.bind(this));
-        return (
-            <div className="col-xs-offset-2 col-xs-8">
-                <ul className="pluginList">
-                    {plugins}
-                </ul>
-            </div>
-        );
-    }
-});
-
 
 var App = React.createClass({
     getInitialState: function() {
@@ -466,6 +223,5 @@ function xhrRequest(url, success, fail) {
     xhr.open("GET", url, true);
     xhr.send();
 }
-
 
 module.exports = window.App = App;
